add system path to find web_scraper.py module in 'logic' folder for use in main.py. Also brought main.py file out of presentation and into root folder.
